refactor: 💡 modified field option children module
This commit will modify field option children module
